make v1,v2 Iterators in DirectDruidClient driven by factory
