+++ server/src/test/java/io/druid/client/V3QueryResponseIteratorTest.java
@@ -1,21 +1,21 @@
 /*
- * Licensed to Metamarkets Group Inc. (Metamarkets) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. Metamarkets licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
+* Licensed to Metamarkets Group Inc. (Metamarkets) under one
+* or more contributor license agreements. See the NOTICE file
+* distributed with this work for additional information
+* regarding copyright ownership. Metamarkets licenses this file
+* to you under the Apache License, Version 2.0 (the
+* "License"); you may not use this file except in compliance
+* with the License. You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing,
+* software distributed under the License is distributed on an
+* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+* KIND, either express or implied. See the License for the
+* specific language governing permissions and limitations
+* under the License.
+*/
 
 package io.druid.client;
 
@@ -42,7 +42,7 @@ import java.util.concurrent.Future;
 
 /**
  */
-public class JsonParserV3ResponseIteratorTest
+public class V3QueryResponseIteratorTest
 {
   @Test
   public void testSimpleResponse() throws Exception
@@ -95,7 +95,7 @@ public class JsonParserV3ResponseIteratorTest
 
     Map<String, Object> context = new HashMap<>();
 
-    DirectDruidClient.JsonParserV3ResponseIterator<String> parser = new DirectDruidClient.JsonParserV3ResponseIterator<>(
+    V3QueryResponseIterator<String> parser = new V3QueryResponseIterator<>(
         baseType,
         responseFuture,
         "http://dummy/",
